매퍼/Consumer 버그 수정 및 메인 Oracle 커넥션을 DPL로 정렬
- ErrorType.notAnalysisTopics()를 static으로 변경 (KafkaListener SpEL의 T(...).method() 호출 호환)
- FarmMapper.xml에 MyBatis DOCTYPE 헤더 추가, VF.FDB_FRMHS_SN 오타 수정, DB Link(@M2MSYS/@DL_LSFARM) 정합
- OracleSqlSessionFactory에 typeAliasesPackage 등록 (resultType 짧은 alias 인식)
- KafkaLivestockErrorEventConsumer에서 currentValue를 Double.parseDouble로 파싱 (운영 메시지 \"63000.0\" 형식 처리)
- application-local.yml의 oracle URL을 DPL PDB로 변경 (매퍼 SQL이 가정하는 메인 connection과 일치)

+++ src/main/java/kr/go/kahis/batchmonitor/messaging/consumer/KafkaLivestockErrorEventConsumer.java
@@ -40,7 +40,8 @@ public class KafkaLivestockErrorEventConsumer implements KafkaConsumer {
 
     try {
       readerService.analysis(event, event.metadata().get("farmNumber"),
-          event.metadata().get("speciesCode"), Long.parseLong(event.metadata().get("currentValue")));
+          event.metadata().get("speciesCode"),
+          (long) Double.parseDouble(event.metadata().get("currentValue")));
     } catch (Exception e) {
       log.error("Unexpected error during livestock analysis: {}", event.eventId(), e);
     } finally {
